reduce code
