+++ architecture-rules/pom.xml
@@ -28,5 +28,9 @@
       <artifactId>archunit</artifactId>
       <version>1.4.1</version>
     </dependency>
+    <dependency>
+      <groupId>com.google.code.findbugs</groupId>
+      <artifactId>jsr305</artifactId>
+    </dependency>
   </dependencies>
 </project>
