--- architecture-rules/src/main/java/com/sap/ai/sdk/architecture/AllowModelInheritance.java
@@ -1,14 +1,0 @@
-package com.sap.ai.sdk.architecture;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Documented
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface AllowModelInheritance {
-  String reason() default "";
-}
