+++ architecture-rules/src/main/java/com/sap/ai/sdk/architecture/InheritGeneratedModel.java
@@ -1,0 +1,103 @@
+package com.sap.ai.sdk.architecture;
+
+import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.classes;
+
+import com.tngtech.archunit.base.DescribedPredicate;
+import com.tngtech.archunit.core.domain.JavaClass;
+import com.tngtech.archunit.core.importer.ClassFileImporter;
+import com.tngtech.archunit.core.importer.ImportOption;
+import com.tngtech.archunit.lang.ArchCondition;
+import com.tngtech.archunit.lang.ConditionEvents;
+import com.tngtech.archunit.lang.SimpleConditionEvent;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import javax.annotation.Nonnull;
+
+/**
+ * Runs the global architecture rule that restricts inheritance from packages ending with {@code
+ * .model}.
+ */
+public final class InheritGeneratedModel {
+  private static final DescribedPredicate<JavaClass> MODEL_PACKAGE =
+      JavaClass.Predicates.resideInAPackage("..model");
+
+  private static final DescribedPredicate<JavaClass> NOT_IN_MODEL_PACKAGE =
+      new DescribedPredicate<>("not residing in a package ending with .model") {
+        @Override
+        public boolean test(final JavaClass input) {
+          return !MODEL_PACKAGE.test(input);
+        }
+      };
+
+  private static final DescribedPredicate<JavaClass> NOT_ALLOWLISTED =
+      new DescribedPredicate<>("not annotated with @AllowModelInheritance") {
+        @Override
+        public boolean test(final JavaClass input) {
+          return !input.isAnnotatedWith(AllowModelInheritance.class);
+        }
+      };
+
+  private InheritGeneratedModel() {}
+
+  /**
+   * Executes the inheritance rule for the package configured via {@code
+   * model.inheritance.basePackage}.
+   *
+   * @param args Unused command line arguments.
+   */
+  public static void main(final String[] args) {
+    final var targetPackage = System.getProperty("model.inheritance.basePackage", "");
+    final var importedClasses =
+        new ClassFileImporter()
+            .withImportOption(ImportOption.Predefined.DO_NOT_INCLUDE_TESTS)
+            .importPackages(targetPackage);
+
+    classes()
+        .that(NOT_IN_MODEL_PACKAGE)
+        .and(NOT_ALLOWLISTED)
+        .should(
+            new ArchCondition<>("not extend or implement .model types") {
+              @Override
+              public void check(final JavaClass item, final ConditionEvents events) {
+                if (targetsModelPackage(item)) {
+                  events.add(
+                      SimpleConditionEvent.violated(
+                          item,
+                          item.getFullName()
+                              + " inherits from a type in a package ending with '.model' while not residing in such a package itself. "
+                              + "Use @AllowModelInheritance(reason = \"...\") for approved exceptions."));
+                }
+              }
+            })
+        .check(importedClasses);
+  }
+
+  private static boolean targetsModelPackage(final JavaClass source) {
+    final var superClass = source.getRawSuperclass();
+    if (superClass.isPresent() && MODEL_PACKAGE.test(superClass.get())) {
+      return true;
+    }
+    return source.getRawInterfaces().stream().anyMatch(MODEL_PACKAGE);
+  }
+
+  /**
+   * Marks approved exceptions to the model inheritance rule.
+   *
+   * <p>Types annotated with this are excluded from the ArchUnit check.
+   */
+  @Documented
+  @Retention(RetentionPolicy.RUNTIME)
+  @Target(ElementType.TYPE)
+  public @interface AllowModelInheritance {
+    /**
+     * Explains why the exception is needed.
+     *
+     * @return The reason for the exception.
+     */
+    @Nonnull
+    String reason() default "";
+  }
+}
